requirements.txt
boto3==1.9.130
botocore==1.12.130
certifi==2019.3.9
chardet==3.0.4
config==0.4.0
docutils==0.14
idna==2.8
jmespath==0.9.4
numpy==1.16.2
pandas==0.24.2
py4j==0.10.7
pyspark==2.4.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
